tests are almost fixed
